// frontend/src/api/generated/model/sellerProfileCreateRequestDto.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */

export interface SellerProfileCreateRequestDto {
  /**
   * @minLength 1
   * @pattern ^[^<>"'\\;]{2,30}$
   */
  storeName: string;
  /** @pattern .*\S.* */
  storeImageUrl?: string;
  /** @pattern .*\S.* */
  snsUrl?: string;
  /** @pattern ^[0-9-]{8,13}$ */
  storePhoneNumber?: string;
  /**
   * @minLength 0
   * @maxLength 100
   */
  storeDescription?: string;
}