// frontend/src/api/generated/model/sellerProfileResponseDto.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */

export interface SellerProfileResponseDto {
  sellerProfileId?: number;
  storeName?: string;
  storeImageUrl?: string;
  snsUrl?: string;
  storePhoneNumber?: string;
  storeDescription?: string;
  createdAt?: string;
}